requirements.txt
tqdm
requests
pillow==10.2.0
pycryptodome
pyzbar
qrcode